requirements.txt
fastapi_utils==0.2.1
fastapi==0.111.0
gunicorn==20.0.4
starlette==0.37.2
sse-starlette==2.1.2
loguru==0.5.3
mysql-connector-python==9.0.0
pandas